Fix the latestreleasedate feature to not write any output unless enabled. Split extension test files.
