+++ kingfisher_scrapy/extensions.py
@@ -19,25 +19,29 @@ class KingfisherLatestDate:
 
     @classmethod
     def from_crawler(cls, crawler):
-        path = crawler.settings['KINGFISHER_LATEST_RELEASE_DATE_FILE_PATH']
-        os.makedirs(path, exist_ok=True)
-        filename = os.path.join(path, 'dates.csv')
+        path = crawler.settings['KINGFISHER_LATEST_RELEASE_DATE_PATH']
+        filename = os.path.join(path, 'latestreleasedate.csv')
+
         extension = cls(filename=filename)
         crawler.signals.connect(extension.item_scraped, signal=signals.item_scraped)
         crawler.signals.connect(extension.spider_closed, signal=signals.spider_closed)
+
         return extension
 
     def item_scraped(self, item, spider):
-        if not isinstance(item, LatestReleaseDateItem) or spider.name in self.spiders_seen:
+        if not spider.latest or spider.name in self.spiders_seen or not isinstance(item, LatestReleaseDateItem):
             return
+
         self.spiders_seen.add(spider.name)
         with open(self.filename, 'a+') as output:
             output.write(f"{item['date']},{spider.name}\n")
 
     def spider_closed(self, spider, reason):
-        if spider.name not in self.spiders_seen:
-            with open(self.filename, 'a+') as output:
-                output.write(f"{reason},{spider.name}\n")
+        if not spider.latest or spider.name in self.spiders_seen:
+            return
+
+        with open(self.filename, 'a+') as output:
+            output.write(f"{reason},{spider.name}\n")
 
 
 class KingfisherFilesStore:
