+++ tests/extensions/test_kingfisher_process_api.py
@@ -1,23 +1,12 @@
-import json
 import os
-from tempfile import TemporaryDirectory
 from unittest.mock import Mock, patch
 
 import pytest
 from scrapy.exceptions import NotConfigured
 
-from kingfisher_scrapy.extensions import KingfisherFilesStore, KingfisherLatestDate, KingfisherProcessAPI
-from kingfisher_scrapy.items import FileError, LatestReleaseDateItem
-from tests import spider_with_crawler
-
-
-def spider_with_files_store(files_store, **kwargs):
-    spider = spider_with_crawler(**kwargs)
-    spider.crawler.settings['FILES_STORE'] = files_store
-    spider.crawler.settings['KINGFISHER_API_URI'] = 'http://httpbin.org/anything'
-    spider.crawler.settings['KINGFISHER_API_KEY'] = 'xxx'
-
-    return spider
+from kingfisher_scrapy.extensions import KingfisherFilesStore, KingfisherProcessAPI
+from kingfisher_scrapy.items import FileError
+from tests import spider_with_crawler, spider_with_files_store
 
 
 def test_from_crawler():
@@ -26,9 +15,9 @@ def test_from_crawler():
     spider.crawler.settings['KINGFISHER_API_KEY'] = 'xxx'
     spider.crawler.settings['KINGFISHER_API_LOCAL_DIRECTORY'] = 'localdir'
 
-    api_extension = KingfisherProcessAPI.from_crawler(spider.crawler)
+    extension = KingfisherProcessAPI.from_crawler(spider.crawler)
 
-    assert api_extension.directory == 'localdir'
+    assert extension.directory == 'localdir'
 
 
 @pytest.mark.parametrize('api_url,api_key', [
@@ -135,7 +124,7 @@ def test_item_scraped_file(sample, is_sample, path, note, encoding, encoding2, d
 def test_item_scraped_file_item(sample, is_sample, note, encoding, encoding2, ok, tmpdir, caplog):
     spider = spider_with_files_store(tmpdir, sample=sample, note=note)
 
-    api_extension = KingfisherProcessAPI.from_crawler(spider.crawler)
+    extension = KingfisherProcessAPI.from_crawler(spider.crawler)
 
     with patch('requests.post') as mocked:
         response = Mock()
@@ -155,7 +144,7 @@ def test_item_scraped_file_item(sample, is_sample, note, encoding, encoding2, ok
             encoding=encoding2,
         )
 
-        api_extension.item_scraped(item, spider)
+        extension.item_scraped(item, spider)
 
         if not ok:
             message = 'Failed to post [https://example.com/remote.json]. API status code: 400'
@@ -198,7 +187,7 @@ def test_item_scraped_file_item(sample, is_sample, note, encoding, encoding2, ok
 def test_item_scraped_file_error(sample, is_sample, ok, tmpdir, caplog):
     spider = spider_with_files_store(tmpdir, sample=sample)
 
-    api_extension = KingfisherProcessAPI.from_crawler(spider.crawler)
+    extension = KingfisherProcessAPI.from_crawler(spider.crawler)
 
     with patch('requests.post') as mocked:
         response = Mock()
@@ -212,7 +201,7 @@ def test_item_scraped_file_error(sample, is_sample, ok, tmpdir, caplog):
             'errors': {'http_code': 500},
         })
 
-        api_extension.item_scraped(data, spider)
+        extension.item_scraped(data, spider)
 
         if not ok:
             message = 'Failed to post [https://example.com/remote.json]. File Errors API status code: 400'
@@ -250,7 +239,7 @@ def test_item_scraped_file_error(sample, is_sample, ok, tmpdir, caplog):
 def test_spider_closed(sample, is_sample, ok, tmpdir, caplog):
     spider = spider_with_files_store(tmpdir, sample=sample)
 
-    api_extension = KingfisherProcessAPI.from_crawler(spider.crawler)
+    extension = KingfisherProcessAPI.from_crawler(spider.crawler)
 
     with patch('requests.post') as mocked:
         response = Mock()
@@ -258,7 +247,7 @@ def test_spider_closed(sample, is_sample, ok, tmpdir, caplog):
         response.status_code = 400
         mocked.return_value = response
 
-        api_extension.spider_closed(spider, 'finished')
+        extension.spider_closed(spider, 'finished')
 
         mocked.assert_called_once_with(
             'http://httpbin.org/anything/api/v1/submit/end_collection_store/',
@@ -286,109 +275,9 @@ def test_spider_closed(sample, is_sample, ok, tmpdir, caplog):
 def test_spider_closed_other_reason(tmpdir):
     spider = spider_with_files_store(tmpdir)
 
-    api_extension = KingfisherProcessAPI.from_crawler(spider.crawler)
+    extension = KingfisherProcessAPI.from_crawler(spider.crawler)
 
     with patch('requests.post') as mocked:
-        api_extension.spider_closed(spider, 'xxx')
+        extension.spider_closed(spider, 'xxx')
 
         mocked.assert_not_called()
-
-
-@pytest.mark.parametrize('sample,path', [
-    (None, os.path.join('test', '20010203_040506', 'file.json')),
-    ('true', os.path.join('test_sample', '20010203_040506', 'file.json')),
-])
-def test_item_scraped_with_build_file_from_response(sample, path, tmpdir):
-    spider = spider_with_files_store(tmpdir, sample=sample)
-    store_extension = KingfisherFilesStore.from_crawler(spider.crawler)
-
-    response = Mock()
-    response.body = b'{"key": "value"}'
-    response.request = Mock()
-    response.request.url = 'https://example.com/remote.json'
-
-    item = spider.build_file_from_response(response, file_name='file.json', data_type='release_package',
-                                           encoding='iso-8859-1')
-    store_extension.item_scraped(item, spider)
-
-    with open(tmpdir.join(path)) as f:
-        assert f.read() == '{"key": "value"}'
-
-    with open(tmpdir.join(path + '.fileinfo')) as f:
-        assert json.load(f) == {
-            'url': 'https://example.com/remote.json',
-            'data_type': 'release_package',
-            'encoding': 'iso-8859-1',
-        }
-
-    assert item['path'] == path
-    assert item['files_store'] == tmpdir
-
-
-@pytest.mark.parametrize('sample,path', [
-    (None, os.path.join('test', '20010203_040506', 'file.json')),
-    ('true', os.path.join('test_sample', '20010203_040506', 'file.json')),
-])
-def test_item_scraped_with_build_file(sample, path, tmpdir):
-    spider = spider_with_files_store(tmpdir, sample=sample)
-    store_extension = KingfisherFilesStore.from_crawler(spider.crawler)
-
-    data = b'{"key": "value"}'
-    url = 'https://example.com/remote.json'
-
-    item = spider.build_file(file_name='file.json', url=url, data=data, data_type='release_package',
-                             encoding='iso-8859-1')
-    store_extension.item_scraped(item, spider)
-
-    with open(tmpdir.join(path)) as f:
-        assert f.read() == '{"key": "value"}'
-
-    with open(tmpdir.join(path + '.fileinfo')) as f:
-        assert json.load(f) == {
-            'url': 'https://example.com/remote.json',
-            'data_type': 'release_package',
-            'encoding': 'iso-8859-1',
-        }
-
-    assert item['path'] == path
-    assert item['files_store'] == tmpdir
-
-
-def test_build_file_with_existing_directory():
-    spider = spider_with_crawler()
-
-    with TemporaryDirectory() as tmpdirname:
-        files_store = os.path.join(tmpdirname, 'data')
-        spider.crawler.settings['FILES_STORE'] = files_store
-        store_extension = KingfisherFilesStore.from_crawler(spider.crawler)
-        os.makedirs(os.path.join(files_store, 'test', '20010203_040506'))
-
-        # No FileExistsError exception.
-        store_extension.item_scraped(spider.build_file(file_name='file.json', data=b'{"key": "value"}'), spider)
-
-
-def test_item_scraped_latest_date():
-    with TemporaryDirectory() as tmpdirname:
-        spider = spider_with_files_store(tmpdirname, latest=True)
-        spider.crawler.settings['KINGFISHER_LATEST_RELEASE_DATE_FILE_PATH'] = tmpdirname
-
-        latest_extension = KingfisherLatestDate.from_crawler(spider.crawler)
-        item = LatestReleaseDateItem({'date': '2020-10-01'})
-        latest_extension.item_scraped(item, spider)
-
-        with open(os.path.join(tmpdirname, 'dates.csv')) as f:
-            assert '2020-10-01,test\n' == f.read()
-
-        # the same item is processed just once
-        latest_extension.item_scraped(item, spider)
-
-        with open(os.path.join(tmpdirname, 'dates.csv')) as f:
-            assert '2020-10-01,test\n' == f.read()
-
-        # a non processed item is marked as an error
-        spider.name = 'dateless'
-
-        latest_extension.spider_closed(spider, 'itemcount')
-
-        with open(os.path.join(tmpdirname, 'dates.csv')) as f:
-            assert '2020-10-01,test\nitemcount,dateless\n' == f.read()
